check for refresh token before login
